blacked and moved rxpower function
