+++ apps/mobility_robustness_optimization/simple_mro.py
@@ -37,7 +37,7 @@ class SimpleMRO(MobilityRobustnessOptimization):
             )
 
         # Generate and preprocess simulation data
-        self.simulation_data = get_ue_data(self.mobility_params)
+        self.simulation_data = get_ue_data(self.mobility_model_params)
         self.simulation_data = self.simulation_data.rename(
             columns={"lat": "latitude", "lon": "longitude"}
         )
